Remove HHVM test target from Travis CI
The default Travis CI image is not able to install it anymore so given
the trend back to PHP 7 instead of HHVM I’ve decided to just drop
official support (backed by tests) for HHVM and move on. In the short
term nothing changes because the code obviously still works on HHVM as
well.